refactored testing and smaller updates
